Refactor Routing
