+++ frontend/src/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Router from './Router'
+import Router from './router/Router'
 
 export default function App() {
   return (
