+++ frontend/src/components/container/sidenav/Sidenav.module.css
@@ -1,17 +1,21 @@
 .Sidenav {
   display: grid;
-  grid-template-columns: 2fr 1fr;
+  grid-template-columns: 3fr 1fr;
 }
 
 .Sidenav nav {
+  max-width: 320px;
   height: 100%;
   background-color: var(--sidenav-bg);
   color: var(--sidenav-text);
-  z-index: 2;
+  z-index: 3;
   font-size: 14px;
   font-weight: 500;
   transform: translateX(-110vw);
   transition: transform 0.3s ease-in-out;
+  height: 100vh;
+  position: sticky;
+  top: 0;
 }
 
 .sidenavOpen nav {
@@ -23,7 +27,7 @@
   position: fixed;
   display: block;
   border: none;
-  z-index: 1;
+  z-index: 2;
   top: 0;
   width: 100vw;
   right: 0;
@@ -31,7 +35,7 @@
   outline: none;
 }
 
-@media (min-width: 768px) {
+@media (min-width: 1024px) {
   .Sidenav {
     display: block;
     height: 100vh;
